requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
apex==0.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
bcrypt==3.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
efficientnet-pytorch==0.6.3
et-xmlfile==1.1.0
Flask==1.1.2
fsspec==2021.4.0
future==0.18.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
multidict==5.1.0
munch==2.5.0
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.2.52
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
paramiko==2.7.2
Pillow==8.2.0
pretrainedmodels==0.7.4
pretty-errors==1.2.22
prettytable==2.1.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.3.3
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.4.1
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@a6e1123983548be55d4d1320e0a2f5fd9174d4ac
Shapely==1.7.1
six==1.16.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
timm==0.3.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchmetrics==0.3.2
torchsummary==1.5.1
torchvision==0.9.1+cu111
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
